Rename seedadmission package to seedadmissioncontroller
