+++ pkg/seedadmissioncontroller/webhooks/admission/extensioncrds/testdata/backupentry.yaml
