+++ pkg/seedadmissioncontroller/webhooks/admission/extensioncrds/testdata/controlplane.yaml
