+++ pkg/seedadmissioncontroller/webhooks/admission/extensioncrds/testdata/extension.yaml
