+++ pkg/seedadmissioncontroller/webhooks/admission/extensioncrds/testdata/network.yaml
