+++ pkg/seedadmissioncontroller/webhooks/admission/extensioncrds/testdata/osc.yaml
