+++ pkg/seedadmissioncontroller/webhooks/admission/extensioncrds/testdata/worker.yaml
